Add files via upload
(Data Cleaning):
One of the most important roles of a data analyst is to clean and make sense of the data
presented to them. So, as a data analyst let's suppose you are given the task of cleaning
the following dataset of Famous Movie Reviews with Python path. Take us through your
data cleaning process and share the final data that you think is ready for use.
Movie dataset:
https://docs.google.com/spreadsheets/d/1nUbyyp021YOclKGNqxdyxG_voAoRjsZtTVbx
0o6WJ-E/edit?usp=sharing